pyproject.toml
[project]
name = "home-photo-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "julianesja",email = "julianesja@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "face-recognition>=1.3.0,<2.0.0",
    "face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models.git",
    "setuptools>=80.9.0,<81.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "mysql-connector-python>=8.1.0,<9.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
    "Pillow>=11.1.0,<12.0.0",
    "pillow-heif>=1.0.0,<2.0.0",
    "imagehash>=4.3.2,<5.0.0",
    "minio>=7.2.16,<8.0.0",
    "qdrant-client<1.15.0",
    "filetype>=1.2.0,<2.0.0",
    "sentence-transformers>=5.1.1,<6.0.0",
    "dependency-injector>=4.48.2,<5.0.0",
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
